refactor: seperate CountdownUntilMigration
